Energy Pylon

+++ src/main/java/com/finderfeed/solarforge/ClientHelpers.java
@@ -38,7 +38,7 @@ public class ClientHelpers {
     }
 
 
-    public static void updateClientRunicEnergyForPlayer(int amount, RunicEnergy.Type type){
+    public static void updateClientRunicEnergyForPlayer(float amount, RunicEnergy.Type type){
         RunicEnergy.setEnergy(getClientPlayer(),amount,type);
     }
